+++ all/be_memosig_triple/index.d.ts
@@ -1,16 +1,16 @@
-import type { Be, be_config_arg_a_T, ctx__be_T, ctx__get_T, ctx__set_T, Ctx_wide_T } from '../be_/index.js'
+import type { Be, be_config_arg_a_T, ctx__be_T, ctx__get_T, ctx__set_T, wide_ctx_T } from '../be_/index.js'
 import type { sig_T } from '../rmemo/index.js'
 export declare function be_memosig_triple_<
 	val_T,
 	ns_T extends string = '',
 	E = unknown,
-	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
+	ctx_T extends wide_ctx_T<ns_T> = wide_ctx_T<ns_T>,
 >(be:Be<sig_T<val_T, E>, ns_T, ctx_T>):be_memosig_triple_T<val_T, ns_T, E, ctx_T>
 export declare function be_memosig_triple_<
 	val_T,
 	ns_T extends string = '',
 	E = unknown,
-	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
+	ctx_T extends wide_ctx_T<ns_T> = wide_ctx_T<ns_T>,
 >(
 	val__new:(ctx:ctx_T, memosig:sig_T<val_T, E>)=>val_T,
 	...config:be_config_arg_a_T<ns_T>
@@ -19,7 +19,7 @@ export declare function ns_be_memosig_triple_<
 	val_T,
 	ns_T extends string,
 	E = unknown,
-	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
+	ctx_T extends wide_ctx_T<ns_T> = wide_ctx_T<ns_T>,
 >(
 	ns:ns_T,
 	val__new:(ctx:ctx_T, memosig:sig_T<val_T, E>)=>val_T,
@@ -27,7 +27,7 @@ export declare function ns_be_memosig_triple_<
 export declare function id_be_memosig_triple_<
 	val_T,
 	E = unknown,
-	ctx_T extends Ctx_wide_T<''> = Ctx_wide_T<''>,
+	ctx_T extends wide_ctx_T<''> = wide_ctx_T<''>,
 >(
 	id:string,
 	val__new:(ctx:ctx_T, memosig:sig_T<val_T, E>)=>val_T,
@@ -36,7 +36,7 @@ export declare function ns_id_be_memosig_triple_<
 	val_T,
 	ns_T extends string,
 	E = unknown,
-	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
+	ctx_T extends wide_ctx_T<ns_T> = wide_ctx_T<ns_T>,
 >(
 	ns:ns_T,
 	id:string,
@@ -46,7 +46,7 @@ export type be_memosig_triple_T<
 	val_T,
 	ns_T extends string = '',
 	E = unknown,
-	ctx_T extends Ctx_wide_T<ns_T> = Ctx_wide_T<ns_T>,
+	ctx_T extends wide_ctx_T<ns_T> = wide_ctx_T<ns_T>,
 > = [
 	ctx__be_T<sig_T<val_T, E>, ns_T, ctx_T>,
 	ctx__get_T<val_T, ns_T, ctx_T>,