+++ all/be_memosig_triple/index.test.ts
@@ -1,6 +1,6 @@
 import { test } from 'uvu'
 import { equal } from 'uvu/assert'
-import { be_, type Ctx_wide_T } from '../be_/index.js'
+import { be_, type wide_ctx_T } from '../be_/index.js'
 import { be_sig_triple_ } from '../be_sig_triple/index.js'
 import { ctx__new, ns_ctx__new } from '../ctx/index.js'
 import { memo_, memosig_, type sig_T } from '../rmemo/index.js'
@@ -20,7 +20,7 @@ test('be_memosig_triple_', ()=>{
 		count:number
 	}>((_ctx, foobar$)=>{
 		/* eslint-disable @typescript-eslint/no-unused-vars */
-		type test__ctx = Expect<Equal<typeof _ctx, Ctx_wide_T<''>>>
+		type test__ctx = Expect<Equal<typeof _ctx, wide_ctx_T<''>>>
 		/* eslint-enable @typescript-eslint/no-unused-vars */
 		equal(_ctx, ctx)
 		foobar$.count = (foobar$.count ||= 0) + 1
@@ -63,14 +63,14 @@ test('be_memosig_triple_|+id|+ns', ()=>{
 	] = be_memosig_triple_<number, 'test_ns'>(
 		ctx=>{
 			/* eslint-disable @typescript-eslint/no-unused-vars */
-			type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+			type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 			/* eslint-enable @typescript-eslint/no-unused-vars */
 			return base_(ctx) + 1
 		},
 		{ id: 'foobar', ns: 'test_ns' }
 	).add((ctx, foobar$)=>memo_(()=>{
 		/* eslint-disable @typescript-eslint/no-unused-vars */
-		type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+		type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 		/* eslint-enable @typescript-eslint/no-unused-vars */
 		add_count++
 		add_dep__set(ctx, add_count + foobar$())
@@ -116,7 +116,7 @@ test('be_memosig_triple_|+be', ()=>{
 	] = be_memosig_triple_<number, 'test_ns', { custom: string }>(
 		be_(ctx=>{
 			/* eslint-disable @typescript-eslint/no-unused-vars */
-			type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+			type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 			/* eslint-enable @typescript-eslint/no-unused-vars */
 			const foobar$ =
 				memosig_<number, { custom: string }>(
@@ -178,13 +178,13 @@ test('ns_be_memosig_triple_', ()=>{
 		'test_ns',
 		ctx=>{
 			/* eslint-disable @typescript-eslint/no-unused-vars */
-			type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+			type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 			/* eslint-enable @typescript-eslint/no-unused-vars */
 			return base_(ctx) + 1
 		}
 	).add((ctx, foobar$)=>memo_(()=>{
 		/* eslint-disable @typescript-eslint/no-unused-vars */
-		type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+		type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 		/* eslint-enable @typescript-eslint/no-unused-vars */
 		add_count++
 		add_dep__set(ctx, add_count + foobar$())
@@ -233,13 +233,13 @@ test('id_be_memosig_triple_', ()=>{
 		'foobar',
 		ctx=>{
 			/* eslint-disable @typescript-eslint/no-unused-vars */
-			type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<''>>>
+			type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<''>>>
 			/* eslint-enable @typescript-eslint/no-unused-vars */
 			return base_(ctx) + 1
 		}
 	).add((ctx, foobar$)=>memo_(()=>{
 		/* eslint-disable @typescript-eslint/no-unused-vars */
-		type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<''>>>
+		type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<''>>>
 		/* eslint-enable @typescript-eslint/no-unused-vars */
 		add_count++
 		add_dep__set(ctx, add_count + foobar$())
@@ -288,13 +288,13 @@ test('ns_id_be_memosig_triple_', ()=>{
 		'foobar',
 		ctx=>{
 			/* eslint-disable @typescript-eslint/no-unused-vars */
-			type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+			type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 			/* eslint-enable @typescript-eslint/no-unused-vars */
 			return base_(ctx) + 1
 		},
 	).add((ctx, foobar$)=>memo_(()=>{
 		/* eslint-disable @typescript-eslint/no-unused-vars */
-		type test_ctx = Expect<Equal<typeof ctx, Ctx_wide_T<'test_ns'>>>
+		type test_ctx = Expect<Equal<typeof ctx, wide_ctx_T<'test_ns'>>>
 		/* eslint-enable @typescript-eslint/no-unused-vars */
 		add_count++
 		add_dep__set(ctx, add_count + foobar$())