+++ all/window__ctx__set/index.d.ts
@@ -1,5 +1,5 @@
-import { type Ctx_wide_T } from '../be_/index.js'
-export declare function window__ctx__set(ctx:Ctx_wide_T<''>):void
+import { type wide_ctx_T } from '../be_/index.js'
+export declare function window__ctx__set(ctx:wide_ctx_T<''>):void
 declare global {
 	interface Window {
 		ctx__get<V>(key:string, ns?:string):V