+++ all/window__ctx__set/index.js
@@ -1,6 +1,6 @@
 /// <reference types="ctx-core" />
 /**
- * @param {Ctx_wide_T<''>}ctx
+ * @param {wide_ctx_T<''>}ctx
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export function window__ctx__set(ctx) {